Use functionality settings to configure mining rewards. Simplify BlockDiffer. Carry is not used. Added BlockRewardCalculator and moved some logic from BlockDiffer there. Moved BlockHeader to own file.
