+++ src/main/scala/com/ltonetwork/state/diffs/BlockDiffer.scala
@@ -2,10 +2,9 @@ package com.ltonetwork.state.diffs
 
 import cats.Monoid
 import cats.implicits._
-import cats.syntax.either.catsSyntaxEitherId
 import com.ltonetwork.account.Address
 import com.ltonetwork.block.Block.CurrentBlockFeePart
-import com.ltonetwork.block.{Block, MicroBlock}
+import com.ltonetwork.block.{Block, BlockRewardCalculator, MicroBlock}
 import com.ltonetwork.features.BlockchainFeatures
 import com.ltonetwork.features.FeatureProvider._
 import com.ltonetwork.metrics.Instrumented
@@ -21,19 +20,6 @@ object BlockDiffer extends ScorexLogging with Instrumented {
   val feeBurnAmt: Long = 0.1.lto
   val feeBurnPct       = 0.5
 
-  def blockReward(settings: FunctionalitySettings, bc: Blockchain): Portfolio =
-    Portfolio(balance = bc.featureActivationHeight(BlockchainFeatures.TokenomicsRedefined)
-      .map(height => 10.lto + 1000 * Math.max(height - bc.height + 25000000, 0)).getOrElse(0)) // TODO make these numbers configurable
-
-  def maybeBurnFee(bc: Blockchain, tx: Transaction): Portfolio = {
-    if (bc.isFeatureActivated(BlockchainFeatures.TokenomicsRedefined, bc.height))
-      Portfolio(balance = (tx.fee * (1 - feeBurnPct)).toLong)
-    else if (bc.isFeatureActivated(BlockchainFeatures.BurnFeeture, bc.height))
-      Portfolio(balance = Math.max(0, tx.fee - feeBurnAmt))
-    else
-      Portfolio(balance = tx.fee)
-  }
-
   def fromBlock[Constraint <: MiningConstraint](settings: FunctionalitySettings,
                                                 blockchain: Blockchain,
                                                 maybePrevBlock: Option[Block],
@@ -41,29 +27,27 @@ object BlockDiffer extends ScorexLogging with Instrumented {
                                                 constraint: Constraint,
                                                 verify: Boolean = true): Either[ValidationError, (Diff, Long, Constraint)] = {
     val stateHeight = blockchain.height
+    val blockGenerator = block.signerData.generator
 
-    lazy val prevBlockFeeDistr: Option[Portfolio] = maybePrevBlock.map(b =>
-      Monoid[Portfolio].combineAll(b.transactionData.map { tx =>
-        val fees = maybeBurnFee(blockchain, tx)
-        fees.minus(fees.multiply(CurrentBlockFeePart))
-      }))
-
-    lazy val currentBlockFeeDistr: Option[Portfolio] = None
+    val initDiff = Diff.portfolio(
+      blockGenerator.toAddress -> Monoid[Portfolio].combine(
+        maybePrevBlock.map(b => BlockRewardCalculator.prevBlockFeeDistr(blockchain, b)).orEmpty, // NG reward for closing block
+        BlockRewardCalculator.blockReward(settings, blockchain),
+      )
+    )
 
     for {
-      _ <- block.signaturesValid()
+      _ <- if (verify) block.signaturesValid() else Right(())
       r <- apply(
         settings,
         blockchain,
         constraint,
+        initDiff,
         maybePrevBlock.map(_.timestamp),
-        block.signerData.generator,
-        prevBlockFeeDistr,
-        currentBlockFeeDistr,
+        blockGenerator,
         block.timestamp,
         block.transactionData,
-        stateHeight + 1 //,
-        // verify
+        stateHeight + 1
       )
     } yield r
   }
@@ -74,8 +58,6 @@ object BlockDiffer extends ScorexLogging with Instrumented {
                                                      micro: MicroBlock,
                                                      timestamp: Long,
                                                      constraint: Constraint
-                                                     //,
-                                                     // verify: Boolean = true
   ): Either[ValidationError, (Diff, Long, Constraint)] = {
     for {
       // microblocks are processed within block which is next after 40-only-block which goes on top of activated height
@@ -84,63 +66,45 @@ object BlockDiffer extends ScorexLogging with Instrumented {
         settings,
         blockchain,
         constraint,
+        Diff.empty,
         prevBlockTimestamp,
         micro.sender,
-        None,
-        None,
         timestamp,
         micro.transactionData,
-        blockchain.height // ,
-        // verify
+        blockchain.height
       )
     } yield r
   }
 
+  private def updateConstraint[Constraint <: MiningConstraint](constraint: Constraint, blockchain: Blockchain, tx: Transaction): Constraint =
+    constraint.put(blockchain, tx).asInstanceOf[Constraint]
+
   private def apply[Constraint <: MiningConstraint](settings: FunctionalitySettings,
                                                     blockchain: Blockchain,
                                                     initConstraint: Constraint,
+                                                    initDiff: Diff,
                                                     prevBlockTimestamp: Option[Long],
                                                     blockGenerator: Address,
-                                                    prevBlockFeeDistr: Option[Portfolio],
-                                                    currentBlockFeeDistr: Option[Portfolio],
                                                     timestamp: Long,
                                                     txs: Seq[Transaction],
-                                                    currentBlockHeight: Int // ,
-                                                    // verify: Boolean
+                                                    currentBlockHeight: Int
   ): Either[ValidationError, (Diff, Long, Constraint)] = {
-    def updateConstraint(constraint: Constraint, blockchain: Blockchain, tx: Transaction): Constraint =
-      constraint.put(blockchain, tx).asInstanceOf[Constraint]
-
     val txDiffer = TransactionDiffer(settings, prevBlockTimestamp, timestamp, currentBlockHeight) _
-    val baseDiff = Diff.empty.copy(portfolios = Map(blockGenerator -> blockReward(settings, blockchain)))
-    val initDiff = baseDiff.combine(Diff.empty.copy(portfolios = Map(blockGenerator -> currentBlockFeeDistr.orElse(prevBlockFeeDistr).orEmpty)))
-    val hasNg    = currentBlockFeeDistr.isEmpty
-
-    def calcNextBlockFee(blockchain: Blockchain, portfolio: Portfolio): (Portfolio, Long) = {
-      val ngPf = if (hasNg) {
-        val curPf  = portfolio.multiply(Block.CurrentBlockFeePart)
-        val nextPf = portfolio.minus(curPf)
-        (curPf, nextPf.balance)
-      } else (portfolio, 0L)
-      ngPf.copy(_2 = 0L) // ????
-    }
 
     txs
       .foldLeft((initDiff, 0L, initConstraint).asRight[ValidationError]) {
         case (r @ Left(_), _) => r
-        case (Right((currDiff, carryFee, currConstraint)), tx) =>
+        case (Right((currDiff, _, currConstraint)), tx) =>
           val updatedBlockchain = composite(blockchain, currDiff, 0)
           val updatedConstraint = updateConstraint(currConstraint, updatedBlockchain, tx)
           if (updatedConstraint.isOverfilled)
             Left(ValidationError.GenericError(s"Limit of txs was reached: $initConstraint -> $updatedConstraint"))
           else
             txDiffer(updatedBlockchain, tx).map { newDiff =>
-              val updatedDiff = currDiff.combine(newDiff)
-              if (hasNg) {
-                val (curBlockFees, nextBlockFee) = calcNextBlockFee(updatedBlockchain, maybeBurnFee(blockchain, tx))
-                val diff                         = updatedDiff.combine(Diff.empty.copy(portfolios = Map(blockGenerator -> curBlockFees)))
-                (diff, carryFee + nextBlockFee, updatedConstraint)
-              } else (updatedDiff, 0L, updatedConstraint)
+              val updatedDiff  = currDiff.combine(newDiff)
+              val curBlockFees = BlockRewardCalculator.rewardedFee(blockchain, tx).multiply(CurrentBlockFeePart)
+              val diff         = updatedDiff.combine(Diff.empty.copy(portfolios = Map(blockGenerator -> curBlockFees)))
+              (diff, 0L, updatedConstraint)
             }
       }
   }
